LayoutTests/imported/w3c/web-platform-tests/media-source/mediasource-seek-beyond-duration-expected.txt


FAIL Test seeking beyond updated media duration. assert_equals: Current time equals duration. expected 1.5 but got 1.8
FAIL Test seeking beyond media duration. assert_equals: Current time equals duration. expected 6.5481179138322 but got 6.549
